Base for controlling switch of hands
