+++ src/components/Hands/HandsController.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const HandsController = ({ secondDegrees, minuteDegrees, hourDegrees, name }) => {
+    // Your component logic here
+
+    return (
+        <>
+        </>
+    );
+};
+
+export default HandsController;
